+++ toppy/errors.py
@@ -9,10 +9,18 @@ __all__ = (
     'ClientResponseError',
     'Forbidden',
     'HTTPException',
+    'MissingExtraRequire',
     'NoTokenSet',
     'RateLimited',
     'Unauthorized'
 )
+
+
+class MissingExtraRequire(Exception):
+    """If a user is missing extra requirements"""
+    def __init__(self, require: str):
+        message = f'Missing extra requirements. Install with `pip install toppy[{require}]`'
+        super().__init__(message)
 
 
 class NoTokenSet(Exception):