+++ toppy/utils.py
@@ -1,8 +1,11 @@
 from __future__ import annotations
 
-from typing import Any, Type, TypeVar, Awaitable, Generic
+from typing import Any, Type, TypeVar, Awaitable, Generic, TYPE_CHECKING, Optional
 
 from aiohttp import web
+
+if TYPE_CHECKING:
+    from .webhook import SQLDatabase
 
 
 __all__ = (
@@ -59,7 +62,7 @@ class AsyncContextManager(Generic[T]):
 
 
 async def run_webhook_server(application: web.Application, site_class: Type[web.BaseSite] = web.TCPSite,
-                             **kwargs) -> web.BaseSite:
+                             connect_db: Optional[SQLDatabase] = None, **kwargs) -> web.BaseSite:
     """
     Run the webhook server created in `create_webhook_server`
 
@@ -70,6 +73,9 @@ async def run_webhook_server(application: web.Application, site_class: Type[web.
     site_class: :class:`aiohttp.web.BaseSite`
         The site for the application. Must have all methods from :class:`aiohttp.web.BaseSite`.
         Defaults to :class:`web.TCPSite`
+    connect_db: Optional[:class:`SQLDatabase`]
+        If you pass the :class:`SQLDatabase` instance from :func:`toppy.webhook.create_webhook_server`
+        it will automatically connection to the database.
     **kwargs:
         The kwargs to pass into ``site_class``.
 
@@ -77,6 +83,9 @@ async def run_webhook_server(application: web.Application, site_class: Type[web.
     --------
     The instance of the site class passed into `site_class`.
     """
+    if connect_db is not None:
+        await connect_db.connect()
+
     runner = web.AppRunner(application)
     await runner.setup()