+++ docs/changelog.rst
@@ -13,12 +13,12 @@ Breaking Changes
 1.5.0
     `DBLClient` and `TopGGClient` have been moved to `toppy.client`.
     `webhook` is now a submodule.
-    `create_webhook_server` now returns `tuple[aiohttp.web.Application, SQLDatabse]`
 
 New Features
 -----------------
 1.5.0
     Base classes used now allow easier implementations of new bot websites.
+    Cache for webhook votes.
 
 Bug Fixes
 -----------------